+++ pkg/networkdevice/profile/profiledefinition/metrics.go
@@ -39,16 +39,6 @@ const (
 	ProfileMetricTypePercent ProfileMetricType = "percent"
 )
 
-// KeyValue used to represent mapping
-// Used for RC compatibility (map to list)
-type KeyValue struct {
-	Key   string `yaml:"key" json:"key"`
-	Value string `yaml:"value" json:"value"`
-}
-
-// KeyValueList is a list of mapping key values
-type KeyValueList []KeyValue
-
 // SymbolConfig holds info for a single symbol/oid
 type SymbolConfig struct {
 	OID  string `yaml:"OID,omitempty" json:"OID,omitempty"`
@@ -157,14 +147,6 @@ func (m *MetricsConfig) IsColumn() bool {
 // IsScalar returns true if the metrics config define scalar metrics
 func (m *MetricsConfig) IsScalar() bool {
 	return m.Symbol.OID != "" && m.Symbol.Name != ""
-}
-
-func (kvl *KeyValueList) ToMap() map[string]string {
-	mapping := make(map[string]string)
-	for _, item := range *kvl {
-		mapping[item.Key] = item.Value
-	}
-	return mapping
 }
 
 // NormalizeMetrics converts legacy syntax to new syntax
